+++ src/components/LgNav.vue
@@ -33,7 +33,7 @@
 </template>
 
 <script>
-import TheLogo from '@/components/graphics/TheLogo'
+import TheLogo from '@/components/TheLogo'
 export default {
   name: 'Navigation',
   components: {
