+++ src/components/SmNav.vue
@@ -62,7 +62,7 @@
 </template>
 
 <script>
-import TheLogo from '@/components/graphics/TheLogo'
+import TheLogo from '@/components/TheLogo'
 export default {
   name: 'Navigation',
   components: {
