+++ src/components/TheFooter.vue
@@ -57,7 +57,7 @@
 </template>
 
 <script>
-import TheLogo from '@/components/graphics/TheLogo'
+import TheLogo from '@/components/TheLogo'
 export default {
   name: 'Footer',
   components: {
