+++ src/components/TheLogo.vue
@@ -6,7 +6,7 @@
             viewBox="0 0 128 31"
             aria-labelledby="title-logo"
             role="img">
-            <title id="title-logo">{{ $t('graphics.logo') }}</title>
+            <title id="title-logo">melsicon</title>
             <g
                 stroke="none"
                 stroke-width="1"
