--- src/components/graphics/illustrations/IllustrationCareers.vue
@@ -1,7 +1,0 @@
-<template>
-    <div class="illustration illustration-careers">
-        <img
-            v-lazy="$t('illustrations.careers.src')"
-            :alt="$t('illustrations.careers.desc')">
-    </div>
-</template>
