--- src/components/graphics/illustrations/IllustrationContact.vue
@@ -1,7 +1,0 @@
-<template>
-    <div class="illustration illustration-contact">
-        <img
-            v-lazy="$t('illustrations.contact.src')"
-            :alt="$t('illustrations.contact.desc')">
-    </div>
-</template>
