--- src/components/graphics/illustrations/IllustrationHero.vue
@@ -1,7 +1,0 @@
-<template>
-    <div class="illustration illustration-hero">
-        <img
-            v-lazy="$t('illustrations.hero.src')"
-            :alt="$t('illustrations.hero.desc')">
-    </div>
-</template>
