--- src/components/graphics/illustrations/IllustrationServices.vue
@@ -1,8 +1,0 @@
-<template>
-    <div class="illustration illustration-services">
-        <img
-            v-lazy="$t('illustrations.services.src')"
-            :alt="$t('illustrations.services.desc')">
-
-    </div>
-</template>
