--- src/components/graphics/illustrations/IllustrationTeam.vue
@@ -1,7 +1,0 @@
-<template>
-    <div class="illustration illustration-team">
-        <img
-            v-lazy="$t('illustrations.team.src')"
-            :alt="$t('illustrations.team.desc')">
-    </div>
-</template>
